Refactor code, add some comment and change readme
