+++ docker/docker-compose.yml
@@ -25,7 +25,7 @@ services:
       timeout: 5s
       retries: 5
   minibankc:
-    image: picher/minibankc:2022.0.5-SNAPSHOT
+    image: picher/minibankc:2022.1.0-SNAPSHOT
     container_name: minibankc
     environment:
       PROFILE: "prod"
