+++ pom.xml
@@ -11,7 +11,7 @@
 
     <groupId>com.example</groupId>
     <artifactId>minibankc</artifactId>
-    <version>2022.0.5-SNAPSHOT</version>
+    <version>2022.1.0-SNAPSHOT</version>
 
     <name>minibankc</name>
     <description>Mini BankC Application</description>
