+++ src/main/java/com/example/minibankc/controller/AccountController.java
@@ -6,8 +6,8 @@ import com.example.minibankc.exception.BadRequestAlertException;
 import com.example.minibankc.exception.CustomerNotFoundException;
 import com.example.minibankc.service.AccountService;
 import com.example.minibankc.util.PaginationUtil;
-import com.example.minibankc.util.apects.Loggable;
-import com.example.minibankc.util.apects.RateController;
+import com.example.minibankc.apects.Loggable;
+import com.example.minibankc.apects.RateController;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.extern.slf4j.Slf4j;
