+++ src/main/java/com/example/minibankc/entity/Customer.java
@@ -13,6 +13,7 @@ import java.util.Set;
  * @version 2022.1.1
  * https://www.linkedin.com/in/mahdisharifi/
  * @since 4/24/22
+ * Save customer data
  */
 
 @Entity
