+++ src/main/java/com/example/minibankc/exception/BadRequestAlertException.java
@@ -1,5 +1,6 @@
 package com.example.minibankc.exception;
 
+import lombok.Getter;
 import org.zalando.problem.AbstractThrowableProblem;
 import org.zalando.problem.Status;
 
@@ -7,6 +8,11 @@ import java.net.URI;
 import java.util.HashMap;
 import java.util.Map;
 
+/**
+ * When request is not valid throw this exception
+ * I extracted this class from JHispter
+ */
+@Getter
 public class BadRequestAlertException extends AbstractThrowableProblem {
 
     private static final long serialVersionUID = 1L;
@@ -23,14 +29,6 @@ public class BadRequestAlertException extends AbstractThrowableProblem {
         super(type, defaultMessage, Status.BAD_REQUEST, null, null, null, getAlertParameters(entityName, errorKey));
         this.entityName = entityName;
         this.errorKey = errorKey;
-    }
-
-    public String getEntityName() {
-        return entityName;
-    }
-
-    public String getErrorKey() {
-        return errorKey;
     }
 
     private static Map<String, Object> getAlertParameters(String entityName, String errorKey) {
