+++ src/main/java/com/example/minibankc/repository/CustomerRepository.java
@@ -13,6 +13,5 @@ import org.springframework.stereotype.Repository;
 /**
  * Spring Data SQL repository for the Customer entity.
  */
-@SuppressWarnings("unused")
 @Repository
 public interface CustomerRepository extends JpaRepository<Customer, Long> {}
