+++ src/main/java/com/example/minibankc/service/impl/AccountServiceImpl.java
@@ -7,9 +7,7 @@ import com.example.minibankc.entity.Customer;
 import com.example.minibankc.exception.AccountNotFoundException;
 import com.example.minibankc.exception.CustomerNotFoundException;
 import com.example.minibankc.mapper.AccountMapper;
-import com.example.minibankc.mapper.CustomerMapper;
 import com.example.minibankc.repository.AccountRepository;
-import com.example.minibankc.repository.AccountTransactionRepository;
 import com.example.minibankc.repository.CustomerRepository;
 import com.example.minibankc.service.AccountService;
 import lombok.extern.slf4j.Slf4j;
@@ -29,19 +27,18 @@ import java.util.Optional;
  * @version 2022.1.1
  * https://www.linkedin.com/in/mahdisharifi/
  * @since 4/27/22
+ * Service of account entity
  */
 @Service
 @Transactional
 @Slf4j
 public class AccountServiceImpl implements AccountService {
 
-    @Autowired
-    private MessageSource messages;
-
     private final CustomerRepository customerRepository;
     private final AccountRepository accountRepository;
-
     private final AccountMapper accountMapper;
+    @Autowired
+    private MessageSource messages;
 
     public AccountServiceImpl(CustomerRepository customerRepository, AccountRepository accountRepository, AccountMapper accountMapper) {
         this.customerRepository = customerRepository;
@@ -52,9 +49,9 @@ public class AccountServiceImpl implements AccountService {
     @Override
     public AccountDto findOne(Long id, String lang) throws AccountNotFoundException {
         log.debug("Request to get Account : {}", id);
-        Optional<Account> accountOptional= accountRepository.findById(id);
-        return accountOptional.map(accountMapper::toDto).orElseThrow(()->
-           new AccountNotFoundException(String.format(messages.getMessage("account.notfound", null, new Locale(lang)),id)));
+        Optional<Account> accountOptional = accountRepository.findById(id);
+        return accountOptional.map(accountMapper::toDto).orElseThrow(() ->
+                new AccountNotFoundException(String.format(messages.getMessage("account.notfound", null, new Locale(lang)), id)));
     }
 
     @Override
@@ -64,14 +61,14 @@ public class AccountServiceImpl implements AccountService {
     }
 
     @Override
-    public AccountDto openAccountForExistingCustomer(long customerId,long initialCredit , String lang) throws CustomerNotFoundException{
+    public AccountDto openAccountForExistingCustomer(long customerId, long initialCredit, String lang) throws CustomerNotFoundException {
         log.debug("Request to open an Account. If initialCredit>0 then a transaction will be added : {}", initialCredit);
-        Optional<Customer> customerOptional= customerRepository.findById(customerId);
-        if(customerOptional.isPresent()) {
-            Customer customer=customerOptional.get();
-            Account account=new Account(customer);
+        Optional<Customer> customerOptional = customerRepository.findById(customerId);
+        if (customerOptional.isPresent()) {
+            Customer customer = customerOptional.get();
+            Account account = new Account(customer);
 
-            if(initialCredit>0) {
+            if (initialCredit > 0) {
                 AccountTransaction accountTransaction = new AccountTransaction(initialCredit);
                 accountTransaction.setReferenceNo(Instant.now().toEpochMilli() / 1000);
                 account.addTransaction(accountTransaction);
@@ -79,8 +76,8 @@ public class AccountServiceImpl implements AccountService {
 
             accountRepository.save(account);
             return accountMapper.toDto(account);
-        }else
-           throw new CustomerNotFoundException(String.format(messages.getMessage("customer.notfound", null, new Locale(lang)),customerId));
+        } else
+            throw new CustomerNotFoundException(String.format(messages.getMessage("customer.notfound", null, new Locale(lang)), customerId));
 
     }
